Kill memory servers manually.
This avoids some of the race conditions with tearing the domain
down and killing the memory server. Instead we just kill it right
after blessing (before launching), and when it comes time to kill
the memory servers, we kill everything that matches the URL (which
of course means only the one).