nova-compute now sets up the iptables
Makes calls out to nova-compute so that it is responsible for
setting up the iptables rules for the instances. This ensures that
modifications to security groups, deletion or any future operation
to these rules made by nova-compute continue to work without our
intervention.

This also prevents of having to handle the changing of the iprules
chain when we migrate an instance between hosts.