Adding a timeout parameter for calls to compute for iptables
If there are many instances being launched on a single host then
the timeout on the nova-compute call to setup the iptables can
timeout. This is because modifications done to the nova-compute
are done serially and with many launches at once this can exceed
the 60 second default timeout.

There is now a flag that can be set to raise this timeout. The
value of this flag will largerly be dependent on the expected
number of instances that are launched on a single hosts at once.