Adding a gc-boot command to the gridcentric extension
This basically mimicks the regular nova-compute boot command with
the exception that it must target a host running a nova-gc
service. Basically its a useful testing / demoing command to
bring up a new instance on a particular host.

This feature will only be accessible by admin users so that general
users cannot by-pass the regular nova-boot sequence and the
scheduler.

If no host is specified then a radnom host running the gridcentric
services will be picked.