Allowing migration to not take a destination host.
Fixes issue #352

If there is no destination host specified for the migration then a
random one will be picked from the list of hosts that are running
the gridcentric service and have it enabled. This allows users to
migration instance off a particular host without needing to know,
or caring, where the virtual machine will end up.