Ensure that a migrating instance cannot be migrated.
A check and lock and been put in place in the manager to prevent
a double migration of an instance. The lock is done by adding a
gc:migrating metadata element that is then removed after migration
is successful.

We also ensure that when the API sets the status of the migration
before returning. This will help clients determine when a migration
fails and when it is complete.

Finally, the API can alos do a small check and report back to the
user if the instance is migrating. There is a small race on this
feedback because two API servers can handle migration request at
the same time. The lock in the manager however will prevent both
migrations will from happenning.