Reconfiguring dnsmasq on the src and dest hosts after migration.
The dnsmasq processes for the multi-host networks attached to the
instance are updated on both the source and destination host. On
the source host the instance's entry to removed from the conf file
and on the destination host it is added. If dnsmasq is not running
on the destination host then it will also be started.