Account for the new ini based nova flagfile format in essex.
This is a fix for VMS-312. When adding our extensions to the nova
flagfile we need take the flagfile format into account. Currently
there are two possible formats: the old --flagname format and the new
ini based format introduced in essex. Both formats are accepted by
essex but the two formats can't be mixed in the same file.