Updating Makefile to build the vms version module
The vms.version module needs to be built before running the unit tests
when the VMS_PATH is specified. If using the installed vms packages
then it is assumed that the vms.version module is installed. When
specifying the VMS_PATH it is assume to be a source directory and
the version module needs to be built .

Change-Id: I90ac7095222e600242b93c297d455c7d3ec22ed2