Adding the root nova log handlers to the vms logger
Fixes issue OS-3

Due to a bug in Essex, the root logger loses all of its hanlders and
as a side effect all non nova-based loggers do not log any messages.
We correct this behaviour for the vms logger when configuring the
vms connection so that it will get redirected to the base nova handlers.

Note this has been fixed in Folsom.

Change-Id: I11b5a9e2a212f1e7fb083d8c9bc63f078c657a54