Add instance locks around the main operations
Fixes issue VMS-369

The main operations of bless, launch, discard and migrate are now
decorated by instance locks. This is a way to ensure that for a single
manager instance (i.e. for a single host) gridcentric operations on
an instance are synchronized.

The instance locks are created on demand when the manager is asked
to perform an operation on the instance. They are removed from a local
cache by a periodic tasks when that instance is no longer associted
with the manager's host.

This should help prevent race conditions in the API layer sending
duplicate, or competing, requests to the manager. At the manager
level they will be synchronized.

Change-Id: I5cf453d3c28e4f25b14ce800624584cd58e96e5d