Raise an error when a blessed instance is deleted.
Fixes OS-89

Extends delete for servers to check if the instance is blessed.
If the instance is blessed then it will raise an exception so that
an error message can be sent back to the user. The actual deleting
of the instance is prevented because the disable_terminate flag
is set.

Change-Id: I7bf4da324875abf41b034ce8ce8ee3500f84986d