Fix kernel header management in agent install.
With recent agent updates, we don't need to bother anymore about missing kernel
headers, the install will succeed.

However, dkms will not pull kernel headers if missing (in ubuntu), and may pull
the wrong kernel headers if the version has upgraded (in both ubuntu and
centos). So add for both distros an install of the right kernel headers, with a
proper human-readable WARNING in case the headers can't be found.

Also fix a verbose error when setting tty capabilities.

Change-Id: I91df5d31b107adc60741eca992a23863a87eb2e3