Adding support for multiple VMS versions
Adding the ability to supoprt multiple versions of VMS by spitting
out the direct interaction from the vmsconn. Now vmsconn primary
role to so configure inputs based on the management stack used,
and then call to the vmsapi which will communicate correctly
to the version of vms installed on the system.

Change-Id: Ia1f0cfde19e45cf51ed5e652d23d9b518c23ec46