Adds support to bless/launch instances booted from volume
For instances that were booted from volume the bless command will
take a volume snapshot of each volume attached to the instance.
During launch these snapshots will be created into new volumes
and attached to the launched instance as they were to the booted
instance.

This mimick the behaviour that OpenStack does when doing an
'image-create' from an instance booted from volume.

Change-Id: Ifb803846ce909c406453e2f6100bde9cc67718f2