Support for migrating floating ips
As part of migration all the floating ips are disassociated from
the source host and then associated on the destination host. This
is done before right before launch because we want the instances
to be accessible via this ips during migration.

Note that this is a bit racey because something may happen to the
floating ip between the disassociation and re-association (deleted,
assigned to difference instance). Also this does not support
Quantum based networking.

Also note that this is a Folsom specific change because Grizzly
offers a Network API to perform this operation that prevents this
race and support Quantum.

Change-Id: I360806dfffb1da8bf19a64b8735942f4e23e0bfa