Elevate context in the manager entry points
The main context authorization happens at the API entry point and
it is assumed by the time the request reaches the manager that it
is allowed. Unfortunately, some database methods check the context
for admin-ness. This is being phased out of OpenStack because
checks are happening at the time the request enters. To get around
these database checks, we simply elevate all contexts entering to
admin.

Change-Id: Ib01b6f7932996eb17ebea733408bebbd02d81e3f