Stay on same dashboard's instances table after actions
The success_url of the workflows was previously set to
the full reverse URL path for the nova instances table,
but this meant that after performing an action in the
syspanel instances table the browser would jump back to
the nova dashboard. Since the success_url can be a module
path or a real URL, setting it to "." causes the browser
to be redirected back to the current page.

Change-Id: Ia0b0709f251c1fc791b732ca21cea1e37ba2ff9a