Use system_metadata to store extra instance attributes.
Nova instances have two forms of metadata stores: metadata and
system_metadata. system_metadata was introduced to provide storage
for compute driver / extensions which seems more approperiate for
out extension. More importantly, when moving over to use
nova-conductor the extension manager will only be able to update
the system_metadata (nova-conductor white-lists fields of an
instance that can be updated, and while system_metadata is one of
them, metadata is not).

This converts the usage of metadata within the extension to use
system_metadata instead. This is a necessary first step to support
nova-conductor.

Note: The API still updates some of the metadata on the instance,
in particular the 'launched_from' and 'blessed_from' metadata. This
is done so that we can search instances based on the values. The
system_metadata is not searchable.

Note: A migration path from stable/Folsom to Grizzly will need to
be done to move some of the metadata values over to system_metadata.

Change-Id: I24ea7368df0415db34aac9b8e48b02b13954999f