A nova virt driver that does special launch for a live-image
A subclass of nova's libvirt driver that handles a special case
if the image has properties that signify it as a live-iamge.
In this case, the driver makes a call out to vmsconn to launch
the instance instead of boot it. Otherwise, if the image
is not a live-image then it will just delegate to core nova
libvirt driver.

If the flavor of the instance does not match the flavor of
the live-image, then an exception is raised and the instance
will not be spawned.

Change-Id: Ie5ebeb5eeee6fe2e817009991d7436771df48aef