Use Scheduler select_hosts(...) to determine host to launch on
The Grizzly scheduler has an rpc method select_hosts(...) that
returns the list of hosts best suited for the request_spec. The
extension now creates a scheduler request spec, and asks for the
best hosts for the placement. With this list of hosts the extension
then sends a message to the cobalt-computes on the host.

This can be expanded to allow scheduler hints / filters to be
added to the command line while launching.

Change-Id: I49c81b3eb1ae7eb2ff947b9ba1a1ab62978c2271