Bumping up the default timeout for compute RPC call to 1 hour
The compute RPC call will block because it it performing an
iptables and it synchronizes that process. This means that when
launchg many instance concurrently on the same host, this call can
potentially exceed the default RPC timeout of 1 minute. We bump it
to 1 hour to handle larger number of concurrent launches by default.

Change-Id: Id50d4c9553adde073df3eef0934fefdd61286a3c