Track the networks attached to a live-image
Fixes OS-134

When booting an instance the user can specify a subset of available
networks to attached to the instance. This information was getting
lost when the instance was blessed into a live-image. As a result
when launching the live-image the default networks were being
attached, which ends up being all networks available to the project,
instead of the subset of networks on the instance that was blessed.

This now tracks what networks were attached to the instance when
being blessed and will only request those networks when launching
the new instance.

Change-Id: I9f58942515f9b2b3caf6e9b36514f400aeb8c6cf