Policy distribution through nova.
Implementing functionality in cobalt to allow distribution of
vmspolicyd policies to all known vms hosts. The client takes a path to
a locally accessible policy file, and the contents of this file is
copied into all the policy definition files on all hosts. The compute
service does so basic validation on the provided definitions before
installing them.

Change-Id: I3506a00b9b65d7df6d0a9f878a2b99d0b5c1bbbf