Do not copy the cobalt system metadata when creating an instance copy
The cobalt related system metatdata is used to determine the source
instance when either launching or blessing. In the bless case, it
looks for the 'launched_from' system metadata, and in the launch
case it looks for the 'blessed_from'. There is the assumption that
and instance will only have one of these values set in its system
metadata.

When doing clone-of-clones there was an issue because both of these
values were being set because values where being copied over. Now
we ensure that neither of these values are copied.

Change-Id: I9919c10fd07263b8dc58a218a96622169ca51055