Ensure that a proper power state is associated with each instance
Newer version of the python-novaclient do some sanity checks on
the power_state of the instance before displaying it to the user.
As a result this field can no longer be empty and must be defined.
When creating a new instance we set its power state to NOSTATE,
and in the launch case this will be reset after launch completed.
In the bless case we remain at NOSTATE.

There is also a fixup within the api __init__ that will ensure
that all undeleted instances within the db have their power_state
set. If it is not set, then it will be set to NOSTATE.

Change-Id: Ie97bca0e2de3e08ece7cc775c14963bb67bc5680