Allow raw image types to be created for launched instances
fixes OS-158

If an instance specifies a ramdisk or kernel then these images are
raw copied into the instances directory instead of having a qcow2
overlay. For some reason, when having the qcow2 over the qemu
process fails and the instance is toast. We need to ensure that we
can have the ramdisk and kernel copied as raw into the instances
directory.

Change-Id: Ic9342181be8f4015322264a692f676958ea1435c