Use vmsctl instead of vms library.
Issue OS-175

Instead of using the vms module directly to perform the bless,
launch, discard, etc. operations, use the vmsctl program. The
problem with using the vms module is that some of these operations
are required to run as the root user. Most OpenStack services do
not run as root, and there is a overall desire to not have the
cobalt-compute service to run as root.

To allow cobalt-compute to run as non-root and still execute the
vms functionality it will now go through the vmsctl program and
execute that as root. In other words, vmsctl is executed by nova's
rootwrap (i.e. sudo) to perform the vms operations.

This is also important for the compute virt driver because
nova-compute is not normally executed as root, but rather as the
'nova' user.

Change-Id: Ie248a4474c91de2987388393e6773d262ff5f075