Catch InstanceNotFound exceptions when checking for deletion
Issue OS-177

The cobatl extension checks to see if an instance is blessed or
blessing whenever a user deletes it. This allows the extension to
prompt the user to use 'discard' to remove a blessed, or blessing,
instance. However this can race with the normal deletion of a
regular instance. This causes the API to return an error even
though the deletion was a success.

This catches and ignores the InstanceNotFound exception that can
be raised if the instance was deleted before the extesion was
called.

Change-Id: Ie7d2dca0dbface42d49ece07790a0925d64c56fa