Import the cobalt_topic config from cobalt.nova.api
The cobalt_topic configuration value is defined in the
cobalt.nova.api module. This module was previously imported and used
by the manager, but was removed because the api accesses the database
and manager should be able to run in a no db environment. Once the
cobalt.nova.api module was removed from the manager, the cobalt_topic
configuration value was also removed.

This imports the configuration value from the cobalt.nova.api
module.

Fixes OS-218

Change-Id: Idf7ac1b4742ba476a89e06db7f4db8e5bf54a27b