OS-240, OS-242: patch threading and drop vms-2.6
In Havana, nova's code uses threading, thus requiring it to be
monkeypatched.  Unfortunately, some of vms's python code doesn't work
if threading is monkeypatched. In particular, blessing doesn't work
with monkeypatched threading since it creates threads. Fortunately,
starting with vms-2.7, we use vmsctl to bless, so we can monkeypatch
threading and make nova happy.

To deal with this problem, we've decided to drop vms-2.6 support
entirely and monkeypatch threading. We could have limited this patch
to Havana+, but, in the interest of keeping cobalt versions in sync
with respect to vms compatibility, we've decided to drop vms-2.6
support across the board for cobalt 1.4.

This patch is minimal with repsect to dropping vms-2.6. In upstream
branches (1.5), we remove all of the vms-2.6 related code.

Change-Id: I995231997d513199edb2c97a97ca66508cc2d4c5