Create the base disk for the stubbed out ephemeral disks
Fixes OS-249

In addition to stubbing out the final ephemeral disk location, the
base disk also needs to be stubbed out. This is because the decision
if the ephemeral disk should be created is based on the existence
of both the final disk and the base disk. In other words, if any of
these are missing then the creation of the ephemeral disk will be
attempted. This caused problems with LVM, since cobalt would have
already created the volume for the ephemeral disk.

There are 2 cases that need to be handle: LVM and regular. In the
LVM case, nova never create the base disk location. In this case,
the LVM will always be created. To handle this situation, cobalt
simply touches the base file location so that it exists. Now that
the base file exists, and cobalt already created the volume, the
rest of the code behaves as expected.

In the regular case, the ephemeral disk is actually used. In this
case cobalt calls the same nova function to create the ephemeral
base disk.

Change-Id: I371eb68788ddddcb76a34184a2b4b24d60b5f7cc