OS-254: Don't create a _base stub on every launch
To avoid download the root image from Glance if it wasn't already in
_base, we tricked _create_image into thinking the root image was a
file that was already in _base. The file already in _base was an empty
file with the instance's uuid as a name. The problem with this
approach was that we weren't deleting the file after calling
_create_image. So, after every launch, there'd be one more file in
_base.

This patch has _create_image avoid download the base image in another
way. Rather than tricking _create_image into looking for a different
_base image, we just trick _create_image into not looking for a _base
image at all.

Conflicts fixed when cherry-picking to 1.5:

	cobalt/nova/extension/vmsconn.py

When cherry-picking to 1.5, took advantage of the opportunity to apply the same
treatment to ephemeral disk, if any. (ALC)

Change-Id: I6594ebeb3307ebd3d90a8fe38544874f3fb6d704