Allowing networks to be specified when launching new instances.
OS-252

The networks attached to the instances launched by cobalt are
inherited from the live-image. These networks are taken from the
original master instance. However, there are many use cases where
a different set of networks would ideally be attached.

This allows the set of networks attached to the launch instance to
be overridden. If no networks are specified, then the previous
behaviour of using the live-image networks is used. Otherwise, all
of the network need to be specified in the override. In other words,
if the live-image has 2 networks then 2 networks need to be specified
in the override.

Change-Id: Id3a3f38327477e6e1c000ead155522c385cda725