Add a blank network info instead of copying original's.
OS-253

In Havana instances require a network_info cache value to be
presence. Previsouly the original instance's value was copied
over and this cause live-images to report having an ip address
that was the same as the master. This changes it so that an empty
network_info cache value is used.

Change-Id: I67da3fea62c48b39e5c949af4dfbf0efc5bc5900