Support Havana volume API changes
OS-262

Havana changed their internal volume API to accept the volume_id
instead of the volume object. Cobalt needed to be updated to pass
in the volume_id instead of the volume, and the additional API call
to retrieve the volume was no longer needed.

The snapshot 'source_type', a new Havana block_device_mapping field,
also needed to be applied to the created snapshot. This field is use
when launching the instance to indicate that a new volume should be
created for the launched instance.

Change-Id: I3113b399f9ece4a22096481892960efddb17e74b