Use instance_type information stored in instance's system_metadata
Fixes OS-250

The instance_type/flavor is not a reliable source of information
because it can be deleted after an instance has been created from
it. However, all of the instance_type/flavor information is stored
in an instance's system_metadata. The core nova-compute process
uses the system_metadata to reconstruct the instance_type/flavor
when required.

Now cobalt-compute also uses the instance's system_metadata.

Change-Id: I541aa30b845cb1e752e439ec267bc89dbe55d9e6