Fixed a merge error where instance_uuid was passed into instance_update
OS-272

In Havana, with instance now being full objects instead of dictionaries,
instance_update was modified to work on the instance_obj. Needed to update
one of the lines merged in that still passed in the instance_uuid instead
of the object to instance_update.

Added more unit tests so that network interaction was better tested and
was able to expose this failing code path.

Change-Id: I9f9408b377c72d9bc6469c8840bbbcb95d8b99be