Check for deleted networks associated with live-images
OS-276

Live-images keep track of the networks that were attached to the
original instance. These networks are attached at launch by default
and can be optionally overridden. Previously there was no check that
these networks still existed in the system. If the network did not
exist, then no network information could be allocated to the instance
causing it to fail at launch time.

This patch now checks for the validity of the networks associated with
the live-image before attempting the launch.

Change-Id: Ifd4e4d951fa9f92d4b31675c1f11e5e58eb4171c