Ensure only desired fields are updated during _instance_update(...)
OS-279

The method _instance_update now clears out any local modifications
that may have been made to the instance object before updating/saving
it. This prevents unwanted updates occuring to the database.

This problem was revealed when the instance's host attribute is
temporarily changed during migration. This change is noted in the
object, and during the instance update the host attribute also
get updated. This was an unintended side-effect.

Change-Id: Ibaea13f9d131261ad4882395de76ea3c14df9988