Trivial update to use scheduler's select_destinations rpc method
The select_hosts method on the scheduler rpc API was deprecrated in
Havana in favour of using the select_destinations method wich returns
a tuple (host, node) list instead of just a list of hosts. In Icehouse
the select_hosts method was completely removed and Cobalt needs to
switch over to using the newer method.

Change-Id: Iffb76f42ae118f4d433d20e1560fba5f3057242e